pyproject.toml
[project]
name = "crowd-transcribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.91",
    "dependency-injector>=4.49.0",
    "fastapi>=0.136.0",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crowd_transcribe"]
